# api/swagger.yaml
basePath: /api/v1
definitions:
...
  common.StatusMessage:
    properties:
      message:
        type: string
    type: object
  role.ProfileRole:
    properties:
      id:
        type: string
      name:
        maxLength: 100
        type: string
      quota:
        type: integer
    required:
    - name
    type: object
...
  user.Profile:
    properties:
      email:
        maxLength: 255
        type: string
      first_name:
        maxLength: 255
        type: string
      id:
        type: string
      last_name:
        maxLength: 255
        type: string
      role:
        $ref: '#/definitions/role.ProfileRole'
      source:
        maxLength: 100
        type: string
      status:
        maxLength: 100
        type: string
    required:
    - email
    type: object
  user.SetEnabled:
    properties:
      enabled:
        type: boolean
      id:
        type: string
    type: object
...
info:
  contact: {}
  description: Go-micro-SAAS is a web application skeleton to quickly build new SAAS.
  title: Go-micro-SAAS API
  version: "0.1"
paths:
...
  /roles/:
    get:
      consumes:
      - application/json
      description: Lists all roles of the application
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: Role list endpoint
  /roles/:id:
    put:
      consumes:
      - application/json
      description: Updates the settings of a role
      parameters:
      - description: ID of the role information to patch
        in: path
        name: id
        required: true
        type: integer
      - description: The new version of the role to use for update
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/role.ProfileRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: Role update endpoint
  /users:
    get:
      consumes:
      - application/json
      description: Lists the users of the application.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: List users endpoint
  /users/:id:
    patch:
      consumes:
      - application/json
      description: Updates the target user
      parameters:
      - description: ID of the user information to patch
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: User update endpoint
    put:
      consumes:
      - application/json
      description: Updates the target user
      parameters:
      - description: ID of the user information to patch
        in: path
        name: id
        required: true
        type: integer
      - description: The new version of the user information to use for update
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.Profile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: User update endpoint
  /users/:id/enabled:
    put:
      consumes:
      - application/json
      description: Updates the target user
      parameters:
      - description: ID of the user information to patch
        in: path
        name: id
        required: true
        type: integer
      - description: Whether the user is enabled to log in and upload photos
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.SetEnabled'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: User enable/disable endpoint
  /users/profile:
    get:
      consumes:
      - application/json
      description: Gets the current logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.Profile'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: Get user profile endpoint
swagger: "2.0"